Rename/move maven default modules to provide versioned alternatives
The module name "jboss.container.maven.default.bash" is now provided by two
modules, at jboss/container/maven/default/{default,36}. The default one is
using Maven 3.5. The '36/' path is using Maven 3.6.

For the OpenJDK image at least, we can force the use of the newer module
by depending upon it in an overrides file and setting the version
requirement accordingly. By default, the "default" module *appears* to
be chosen by Cekit's module resolver, so the older images get the right
one automatically. If this was found to not be true in some
circumstances the old module could be forced using the same technique
as the new one.